requirements.txt
paho-mqtt
flask
flask-restful
flasgger